Introduce UI test for MainActivity
By introducing a TestQuoteApplication, and making the test runner
use it instead of the production QuoteApplication class, it is
possible to configure Dagger to use a different set of dependencies for
Android connected tests than the ones used for production.

TestApplicationComponent replaces the ApplicationComponent used in pro-
duction code and allows the connected tests to use the TestApplication-
Module instead.

The TestApplicationModule has a mock implementation of the QuoteData-
Source interface, which can be configured before a test runs to allow
different results to be tested.